backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# API and HTTP
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0

# Environment and configuration
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1

# Async support
asyncio==3.4.3

# JSON handling
orjson==3.9.10

# Development tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging
loguru==0.7.2